[MIG] website_sale_coupon_page: Renamed to website_sale_loyalty_page + Migration to 16.0
TT44378
