--- setup/website_sale_coupon_page/odoo/addons/website_sale_coupon_page
@@ -1,1 +1,0 @@
-../../../../website_sale_coupon_page
