+++ setup/website_sale_loyalty_page/odoo/addons/website_sale_loyalty_page
@@ -1,0 +1,1 @@
+../../../../website_sale_loyalty_page
