+++ setup/website_sale_loyalty_page/setup.py
