+++ website_sale_loyalty_page/controllers/main.py
@@ -8,7 +8,7 @@ class WebsiteSale(http.Controller):
     @http.route(["""/promotions"""], type="http", auth="public", website=True)
     def promotion(self, **post):
         all_promos = (
-            request.env["coupon.program"]
+            request.env["loyalty.program"]
             .sudo()
             .search(
                 [
@@ -21,12 +21,11 @@ class WebsiteSale(http.Controller):
         )
         values = {"promos": []}
         for promo in all_promos:
-            if promo._is_valid_partner(request.env.user.partner_id):
-                values["promos"].append(
-                    {
-                        "id": promo.id,
-                        "image_1920": promo.image_1920,
-                        "public_name": promo.public_name,
-                    }
-                )
-        return request.render("website_sale_coupon_page.promotion_layout", values)
+            values["promos"].append(
+                {
+                    "id": promo.id,
+                    "image_1920": promo.image_1920,
+                    "public_name": promo.public_name,
+                }
+            )
+        return request.render("website_sale_loyalty_page.promotion_layout", values)
