+++ website_sale_loyalty_page/i18n/es.po
@@ -1,6 +1,6 @@
 # Translation of Odoo Server.
 # This file contains the translation of the following modules:
-# 	* website_sale_coupon_page
+# 	* website_sale_loyalty_page
 #
 msgid ""
 msgstr ""
@@ -17,90 +17,90 @@ msgstr ""
 "Plural-Forms: nplurals=2; plural=(n != 1);\n"
 "X-Generator: Poedit 2.0.6\n"
 
-#. module: website_sale_coupon_page
-#: model_terms:ir.ui.view,arch_db:website_sale_coupon_page.available_promotions
+#. module: website_sale_loyalty_page
+#: model_terms:ir.ui.view,arch_db:website_sale_loyalty_page.available_promotions
 msgid "<span>No promotions available</span>"
 msgstr "<span>No hay promociones disponibles</span>"
 
-#. module: website_sale_coupon_page
-#: model:ir.model.fields,field_description:website_sale_coupon_page.field_coupon_program__can_publish
+#. module: website_sale_loyalty_page
+#: model:ir.model.fields,field_description:website_sale_loyalty_page.field_coupon_program__can_publish
 msgid "Can Publish"
 msgstr ""
 
-#. module: website_sale_coupon_page
-#: model_terms:ir.ui.view,arch_db:website_sale_coupon_page.promotion_item
+#. module: website_sale_loyalty_page
+#: model_terms:ir.ui.view,arch_db:website_sale_loyalty_page.promotion_item
 msgid "Close"
 msgstr "Cerrar"
 
-#. module: website_sale_coupon_page
-#: model:ir.model,name:website_sale_coupon_page.model_coupon_program
-msgid "Coupon display on a website"
+#. module: website_sale_loyalty_page
+#: model:ir.model,name:website_sale_loyalty_page.model_coupon_program
+msgid "Loyalty display on a website"
 msgstr ""
 
-#. module: website_sale_coupon_page
-#: model:ir.model.fields,field_description:website_sale_coupon_page.field_coupon_program__image_1920
+#. module: website_sale_loyalty_page
+#: model:ir.model.fields,field_description:website_sale_loyalty_page.field_coupon_program__image_1920
 msgid "Image"
 msgstr ""
 
-#. module: website_sale_coupon_page
-#: model:ir.model.fields,field_description:website_sale_coupon_page.field_coupon_program__image_1024
+#. module: website_sale_loyalty_page
+#: model:ir.model.fields,field_description:website_sale_loyalty_page.field_coupon_program__image_1024
 msgid "Image 1024"
 msgstr ""
 
-#. module: website_sale_coupon_page
-#: model:ir.model.fields,field_description:website_sale_coupon_page.field_coupon_program__image_128
+#. module: website_sale_loyalty_page
+#: model:ir.model.fields,field_description:website_sale_loyalty_page.field_coupon_program__image_128
 msgid "Image 128"
 msgstr ""
 
-#. module: website_sale_coupon_page
-#: model:ir.model.fields,field_description:website_sale_coupon_page.field_coupon_program__image_256
+#. module: website_sale_loyalty_page
+#: model:ir.model.fields,field_description:website_sale_loyalty_page.field_coupon_program__image_256
 msgid "Image 256"
 msgstr ""
 
-#. module: website_sale_coupon_page
-#: model:ir.model.fields,field_description:website_sale_coupon_page.field_coupon_program__image_512
+#. module: website_sale_loyalty_page
+#: model:ir.model.fields,field_description:website_sale_loyalty_page.field_coupon_program__image_512
 msgid "Image 512"
 msgstr ""
 
-#. module: website_sale_coupon_page
-#: model:ir.model.fields,field_description:website_sale_coupon_page.field_coupon_program__is_published
+#. module: website_sale_loyalty_page
+#: model:ir.model.fields,field_description:website_sale_loyalty_page.field_coupon_program__is_published
 msgid "Is Published"
 msgstr ""
 
-#. module: website_sale_coupon_page
-#: model:ir.model.fields,help:website_sale_coupon_page.field_coupon_program__public_name
+#. module: website_sale_loyalty_page
+#: model:ir.model.fields,help:website_sale_loyalty_page.field_coupon_program__public_name
 msgid "Name of the promo showed on website bellow the banner image."
 msgstr ""
 "Nombre de la promoción mostrado en el sitio web debajo del la imagen "
 "promocional."
 
-#. module: website_sale_coupon_page
-#: model_terms:ir.ui.view,arch_db:website_sale_coupon_page.promotion_layout
+#. module: website_sale_loyalty_page
+#: model_terms:ir.ui.view,arch_db:website_sale_loyalty_page.promotion_layout
 msgid "Promotions"
 msgstr "Cupones y Promociones"
 
-#. module: website_sale_coupon_page
-#: model:ir.model.fields,field_description:website_sale_coupon_page.field_coupon_program__public_name
+#. module: website_sale_loyalty_page
+#: model:ir.model.fields,field_description:website_sale_loyalty_page.field_coupon_program__public_name
 msgid "Public Name"
 msgstr "Nombre público"
 
-#. module: website_sale_coupon_page
-#: model:ir.model.fields,help:website_sale_coupon_page.field_coupon_program__website_url
+#. module: website_sale_loyalty_page
+#: model:ir.model.fields,help:website_sale_loyalty_page.field_coupon_program__website_url
 msgid "The full URL to access the document through the website."
 msgstr ""
 
-#. module: website_sale_coupon_page
-#: model:ir.model.fields,field_description:website_sale_coupon_page.field_coupon_program__website_published
+#. module: website_sale_loyalty_page
+#: model:ir.model.fields,field_description:website_sale_loyalty_page.field_coupon_program__website_published
 msgid "Visible on current website"
 msgstr ""
 
-#. module: website_sale_coupon_page
-#: model_terms:ir.ui.view,arch_db:website_sale_coupon_page.coupon_program_view_form_common
+#. module: website_sale_loyalty_page
+#: model_terms:ir.ui.view,arch_db:website_sale_loyalty_page.coupon_program_view_form_common
 msgid "Website"
 msgstr "Sitio Web"
 
-#. module: website_sale_coupon_page
-#: model:ir.model.fields,field_description:website_sale_coupon_page.field_coupon_program__website_url
+#. module: website_sale_loyalty_page
+#: model:ir.model.fields,field_description:website_sale_loyalty_page.field_coupon_program__website_url
 msgid "Website URL"
 msgstr ""
