+++ website_sale_loyalty_page/models/__init__.py
@@ -1,1 +1,1 @@
-from . import coupon_program
+from . import loyalty_program
