+++ website_sale_loyalty_page/readme/CONTRIBUTORS.rst
@@ -3,3 +3,4 @@
   * Pedro M. Baeza
   * Carlos Roca
   * Stefan Ungureanu
+  * Pilar Vargas
