+++ website_sale_loyalty_page/static/description/index.html
@@ -3,8 +3,8 @@
 <html xmlns="http://www.w3.org/1999/xhtml" xml:lang="en" lang="en">
 <head>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
-<meta name="generator" content="Docutils 0.15.1: http://docutils.sourceforge.net/" />
-<title>Website Sale Coupon Page</title>
+<meta name="generator" content="Docutils: http://docutils.sourceforge.net/" />
+<title>Website Sale Loyalty Page</title>
 <style type="text/css">
 
 /*
@@ -360,14 +360,14 @@ ul.auto-toc {
 </style>
 </head>
 <body>
-<div class="document" id="website-sale-coupon-page">
-<h1 class="title">Website Sale Coupon Page</h1>
+<div class="document" id="website-sale-loyalty-page">
+<h1 class="title">Website Sale Loyalty Page</h1>
 
 <!-- !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 !! This file is generated by oca-gen-addon-readme !!
 !! changes will be overwritten.                   !!
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! -->
-<p><a class="reference external" href="https://odoo-community.org/page/development-status"><img alt="Beta" src="https://img.shields.io/badge/maturity-Beta-yellow.png" /></a> <a class="reference external" href="http://www.gnu.org/licenses/lgpl-3.0-standalone.html"><img alt="License: LGPL-3" src="https://img.shields.io/badge/licence-LGPL--3-blue.png" /></a> <a class="reference external" href="https://github.com/OCA/sale-promotion/tree/15.0/website_sale_coupon_page"><img alt="OCA/sale-promotion" src="https://img.shields.io/badge/github-OCA%2Fsale--promotion-lightgray.png?logo=github" /></a> <a class="reference external" href="https://translation.odoo-community.org/projects/sale-promotion-15-0/sale-promotion-15-0-website_sale_coupon_page"><img alt="Translate me on Weblate" src="https://img.shields.io/badge/weblate-Translate%20me-F47D42.png" /></a> <a class="reference external" href="https://runbot.odoo-community.org/runbot/296/15.0"><img alt="Try me on Runbot" src="https://img.shields.io/badge/runbot-Try%20me-875A7B.png" /></a></p>
+<p><a class="reference external" href="https://odoo-community.org/page/development-status"><img alt="Beta" src="https://img.shields.io/badge/maturity-Beta-yellow.png" /></a> <a class="reference external" href="http://www.gnu.org/licenses/lgpl-3.0-standalone.html"><img alt="License: LGPL-3" src="https://img.shields.io/badge/licence-LGPL--3-blue.png" /></a> <a class="reference external" href="https://github.com/OCA/sale-promotion/tree/16.0/website_sale_loyalty_page"><img alt="OCA/sale-promotion" src="https://img.shields.io/badge/github-OCA%2Fsale--promotion-lightgray.png?logo=github" /></a> <a class="reference external" href="https://translation.odoo-community.org/projects/sale-promotion-16-0/sale-promotion-16-0-website_sale_loyalty_page"><img alt="Translate me on Weblate" src="https://img.shields.io/badge/weblate-Translate%20me-F47D42.png" /></a> <a class="reference external" href="https://runboat.odoo-community.org/webui/builds.html?repo=OCA/sale-promotion&amp;target_branch=16.0"><img alt="Try me on Runboat" src="https://img.shields.io/badge/runboat-Try%20me-875A7B.png" /></a></p>
 <p>This module allows to publish promotions on Website using banners uploaded to each one.</p>
 <p><strong>Table of contents</strong></p>
 <div class="contents local topic" id="contents">
@@ -387,10 +387,10 @@ ul.auto-toc {
 <h1><a class="toc-backref" href="#id1">Configuration</a></h1>
 <p>To show the promotions/coupons on website you need to:</p>
 <ol class="arabic simple">
-<li>Go to <em>Sales &gt; Products &gt; Promotion Programs/Coupon Programs</em>.</li>
+<li>Go to <em>Website &gt; eCommerce &gt; Loyalty &gt; Discount &amp; Loyalty</em>.</li>
 <li>Select the program that we will publish.</li>
-<li>In the Website section, upload an image, add a text or both.</li>
-<li>Press on the smart button for publishing the promotion/coupon.</li>
+<li>In the Website tab, upload an image, add a text or both.</li>
+<li>Press on the smart button for publishing the loyalty.</li>
 </ol>
 <p><strong>Note:</strong> The order of the items is defined by a sequence, but the promotions will
 appear before the coupons.</p>
@@ -406,7 +406,7 @@ you will see all published items.</p>
 <p>Bugs are tracked on <a class="reference external" href="https://github.com/OCA/sale-promotion/issues">GitHub Issues</a>.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us smashing it by providing a detailed and welcomed
-<a class="reference external" href="https://github.com/OCA/sale-promotion/issues/new?body=module:%20website_sale_coupon_page%0Aversion:%2015.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**">feedback</a>.</p>
+<a class="reference external" href="https://github.com/OCA/sale-promotion/issues/new?body=module:%20website_sale_loyalty_page%0Aversion:%2016.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**">feedback</a>.</p>
 <p>Do not contact contributors directly about support or help with technical issues.</p>
 </div>
 <div class="section" id="credits">
@@ -424,6 +424,7 @@ If you spotted it first, help us smashing it by providing a detailed and welcome
 <li>Pedro M. Baeza</li>
 <li>Carlos Roca</li>
 <li>Stefan Ungureanu</li>
+<li>Pilar Vargas</li>
 </ul>
 </li>
 </ul>
@@ -435,7 +436,7 @@ If you spotted it first, help us smashing it by providing a detailed and welcome
 <p>OCA, or the Odoo Community Association, is a nonprofit organization whose
 mission is to support the collaborative development of Odoo features and
 promote its widespread use.</p>
-<p>This module is part of the <a class="reference external" href="https://github.com/OCA/sale-promotion/tree/15.0/website_sale_coupon_page">OCA/sale-promotion</a> project on GitHub.</p>
+<p>This module is part of the <a class="reference external" href="https://github.com/OCA/sale-promotion/tree/16.0/website_sale_loyalty_page">OCA/sale-promotion</a> project on GitHub.</p>
 <p>You are welcome to contribute. To learn how please visit <a class="reference external" href="https://odoo-community.org/page/Contribute">https://odoo-community.org/page/Contribute</a>.</p>
 </div>
 </div>
