+++ website_sale_loyalty_page/templates/promotion_templates.xml
@@ -4,7 +4,7 @@
         <t t-call="website.layout">
             <div class="container">
                 <h2 class="mt-2 mb-3">Promotions</h2>
-                <t t-call="website_sale_coupon_page.available_promotions" />
+                <t t-call="website_sale_loyalty_page.available_promotions" />
             </div>
         </t>
     </template>
@@ -13,7 +13,7 @@
         <t t-if="promos">
             <div class="row">
                 <t t-foreach="promos" t-as="promo">
-                    <t t-call="website_sale_coupon_page.promotion_item" />
+                    <t t-call="website_sale_loyalty_page.promotion_item" />
                 </t>
             </div>
         </t>
@@ -32,12 +32,12 @@
             <div class="card" style="width: 18rem;">
                 <a
                     href="#"
-                    data-toggle="modal"
-                    t-att-data-target="'#imagemodal_%s' % promo['id']"
+                    data-bs-toggle="modal"
+                    t-att-data-bs-target="'#imagemodal_%s' % promo['id']"
                 >
                     <img
                         class="card-img-top img-fluid"
-                        t-attf-src="/web/image/coupon.program/#{promo['id']}/image_1920"
+                        t-attf-src="/web/image/loyalty.program/#{promo['id']}/image_1920"
                         t-att-alt="'%s' % promo['id']"
                         t-if="promo['image_1920']"
                     />
@@ -50,7 +50,6 @@
                 class="modal fade"
                 t-att-id="'imagemodal_%s' % promo['id']"
                 tabindex="-1"
-                role="dialog"
                 aria-labelledby="Modal-image"
                 aria-hidden="true"
                 t-if="promo['image_1920']"
@@ -58,20 +57,17 @@
                 <div class="modal-dialog modal-lg modal-dialog-centered">
                     <div class="modal-content">
                         <div class="modal-header">
+                            <h5 class="modal-title pt-3" t-out="promo['public_name']" />
                             <button
                                 type="button"
-                                class="close"
-                                data-dismiss="modal"
+                                class="btn-close"
+                                data-bs-dismiss="modal"
                                 aria-label="Close"
-                            >
-                                <span>
-                                    <i class="fa fa-times" />
-                                </span>
-                            </button>
+                            />
                         </div>
-                        <div class="modal-body">
+                        <div class="modal-body text-center">
                             <img
-                                t-attf-src="/web/image/coupon.program/#{promo['id']}/image_1920"
+                                t-attf-src="/web/image/loyalty.program/#{promo['id']}/image_1920"
                                 t-att-alt="'%s' % promo['id']"
                                 class="img-fluid"
                             />
