--- website_sale_loyalty_page/views/sale_coupon_program_views.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-    <record model="ir.ui.view" id="coupon_program_view_form_common">
-        <field name="model">coupon.program</field>
-        <field name="inherit_id" ref="coupon.coupon_program_view_form_common" />
-        <field name="arch" type="xml">
-            <xpath expr="//div[hasclass('oe_button_box')]" position="inside">
-                <button
-                    name="website_publish_button"
-                    type="object"
-                    class="oe_stat_button"
-                    icon="fa-globe"
-                >
-                    <field name="is_published" widget="website_publish_button" />
-                </button>
-            </xpath>
-            <xpath expr="//group[@name='reward']" position="after">
-                <group name="website" string="Website">
-                    <field name="public_name" />
-                    <field name="image_1920" widget="image" />
-                </group>
-            </xpath>
-        </field>
-    </record>
-</odoo>
